Set dns_domain_name in env var
With the intention to reduce hardcoded domain name "lago.local", set it
in run_suite.sh/lagofy.sh and remove the hard-coding in HE lagoinitfile.

Future patches might remove it elsewhere and use the env var.

Doing it this way because I failed to find a simple way to get from lago
the conf value dns_domain_name.

Change-Id: Ib5a11d4ab23da5519deba3ab7fbd40b7ca8ae4ea
Signed-off-by: Yedidyah Bar David <didi@redhat.com>