ost_utils: Add 'ansible_inventory' method to backend
Whatever backend we decide to use, it always should be responsible
for providing the ansible inventory, as it's a starting point
for the suite.

This patch adds an 'ansible_inventory' interface to 'BaseBackend'
and a concrete implementation for 'LagoBackend'.

Unfortunately we still can't get rid of the ansible inventory
generation that we do in the shell scripts [1]. This will happen
in subsequent patches, when we'll expand the 'backend' module
even more.

[1] https://github.com/oVirt/ovirt-system-tests/blob/e26b69e37f8a028a4444894b81f6511f1333fe8f/run_suite.sh#L235

Change-Id: I0d7b85fce130fc0a8ed0e09265d4bc630b221576
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>