Do not fail if host is rebooting
This is not needed in basic-suite because test_verify_add_hosts waits
until at least one host is up, which takes time, so by the time we
test_verify_add_all_hosts, also the other host is hopefully up. It still
might happen that we fail also here, if we do something that makes
installation of host-1 take significantly more time than that of host-0,
so that it is in 'reboot' state when test_verify_add_all_hosts.

In hosted-engine suites, test_verify_add_hosts does not take time -
because host-0 is already up at this point. So by the time we
test_verify_add_all_hosts, host-1 might be in 'reboot' state.

Change-Id: I3a69948cc8538f6b0a6b57556f0ca2b4431e52b9
Signed-off-by: Yedidyah Bar David <didi@redhat.com>