HE: Run a loop of 'dig'
To test CI's network reliability.

Recently several HE OST runs failed, because ovirt-ha-broker failed in
its network monitor (which uses dig similarly to current patch) and this
caused ha-agent to stop the engine VM. E.g. [1].

This patch is not intended to be merged, right now, but in principle
there is no harm and some value in merging it.

[1] https://jenkins.ovirt.org/job/ovirt-system-tests_he-basic-suite-master/2080/

Change-Id: I89769ed04de9e6fb7ac7b33ba835b5ae4b72b9ac
Signed-off-by: Yedidyah Bar David <didi@redhat.com>